Revert protection on prompts POST
I forgot I was using this endpoint already and protecting it
will break the process when it runs next. I couldv just fix
the processes to send an auth token but the auth stuff isn't
done yet so I really can't so away the decorator goes.
I'll just need to add it back when the auth stuff is done