Add simple API-compatible interface to help remove outdated `records` library
This is needed as `records` is an upgrade blocker, and I would like to
possibly get sqlalchemy, flask, and friends pulled up to
the latest versions without waiting for the full v2 rewrite.
This is one of the steps needed to do that. Actual use
of this will come later